move the Move class into its own file
