+++ ai.py
@@ -199,19 +199,3 @@ class AI:
                     break
             return best_score
 
-
-class Move:
-
-    def __init__(self, xfrom, yfrom, xto, yto, castling_move):
-        self.xfrom = xfrom
-        self.yfrom = yfrom
-        self.xto = xto
-        self.yto = yto
-        self.castling_move = castling_move
-
-    # Returns true iff (xfrom,yfrom) and (xto,yto) are the same.
-    def equals(self, other_move):
-        return self.xfrom == other_move.xfrom and self.yfrom == other_move.yfrom and self.xto == other_move.xto and self.yto == other_move.yto
-
-    def to_string(self):
-        return "(" + str(self.xfrom) + ", " + str(self.yfrom) + ") -> (" + str(self.xto) + ", " + str(self.yto) + ")"
